v1.3.0

+++ src/Models/ContextCoverage.php
@@ -7,10 +7,10 @@ use YandexDirectSDK\Components\Model;
 /** 
  * Class ContextCoverage 
  * 
- * @property-read     integer     $probability
+ * @property-read     double      $probability
  * @property-read     integer     $price
  *                                
- * @method            integer     getProbability()
+ * @method            double      getProbability()
  * @method            integer     getPrice()
  * 
  * @package YandexDirectSDK\Models 
@@ -20,7 +20,7 @@ class ContextCoverage extends Model
     protected static $compatibleCollection = ContextCoverages::class;
 
     protected static $properties = [
-        'probability' => 'integer',
+        'probability' => 'double',
         'price' => 'integer'
     ];
